+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.1.1 (2023-01-06)
+
+### Feat
+
+- **managers/selenium**: adding webdriver-manager to reduce complexity
+- **managers/selenium**: small manager to get a chromium and load a web
+- **core**: adding pr body template
+
 ## v0.1.0 (2023-01-06)
 
 ### Feat